pyproject.toml
[tool.poetry]
name = "arxiv2arxode"
version = "0.1.0"
description = "Undergrad research project"
authors = ["Abdulla <abdullabloushi@gmail.com>"]
license = "UNLICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10, <3.12"
openai = "^1.3.4"
pydantic = "^2.5.2"
isort = "^5.12.0"
langchain = "^0.0.341"
nest-asyncio = "^1.5.8"
chromadb = "^0.4.18"
diskcache = "^5.6.3"
pyautogen = { extras = ["retrievechat"], version = "^0.2.2" }
pypdf = "<=4.0.1"
pgvector = "^0.2.4"
arxiv = "^2.1.0"
urllib3 = "<2.0"
psycopg2-binary = "^2.9.9"
unstructured = {extras = ["pdf"], version = "^0.12.6"}
radon = "^6.0.1"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
